// cosmowcl/src/app/shared/models/api/cosmopratiche/variabileDiFiltro.d.ts
/*
 * Copyright CSI-Piemonte - 2023
 * SPDX-License-Identifier: GPL-3.0-or-later
 */
/**
 * Pratiche
 * API per la gestione pratiche 
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TipoFiltro } from './tipoFiltro.d';
import { FormatoVariabileDiFiltro } from './formatoVariabileDiFiltro.d';
import { TipoPratica } from './tipoPratica.d';


export interface VariabileDiFiltro { 
    id?: number;
    nome: string;
    label: string;
    formato: FormatoVariabileDiFiltro;
    tipoPratica: TipoPratica;
    tipoFiltro: TipoFiltro;
    aggiungereARisultatoRicerca?: boolean;
}